[refactor] move Find1P1CPackage to txdownload
Move-only.
